+++ Source/WinCopies.Util.Desktop.Shared/Data/Converters/MultiConverterBase/MultiConverterBase2.cs
@@ -18,7 +18,6 @@
 using System;
 using System.Globalization;
 using System.Windows.Data;
-using System.Windows.Markup;
 
 using WinCopies.Collections;
 using WinCopies.Collections.DotNetFix.Generic;
@@ -36,40 +35,6 @@ using static WinCopies.Util.Util;
 
 namespace WinCopies.Util.Data
 {
-    /// <summary>
-    /// Provides a base-class for any data-<see cref="MultiBinding"/> converter.
-    /// </summary>
-    [MarkupExtensionReturnType(typeof(IMultiValueConverter))]
-    public abstract class MultiConverterBase : MarkupExtension, IMultiValueConverter
-    {
-        /// <summary>
-        /// Converts source values to a value for the binding target. The data binding engine calls this method when it propagates the values from source bindings to the binding target.
-        /// </summary>
-        /// <param name="values">The array of values that the source bindings in the <see cref="MultiBinding"/> produces. The value <see cref="System.Windows.DependencyProperty.UnsetValue"/> indicates that the source binding has no value to provide for conversion.</param>
-        /// <param name="targetType">The type of the binding target property.</param>
-        /// <param name="parameter">The converter parameter to use.</param>
-        /// <param name="culture">The culture to use in the converter.</param>
-        /// <returns>A converted value.</returns>
-        public abstract object Convert(object[] values, Type targetType, object parameter, CultureInfo culture);
-
-        /// <summary>
-        /// Converts a binding target value to the source binding values.
-        /// </summary>
-        /// <param name="value">The value that the binding target produces.</param>
-        /// <param name="targetTypes">The array of types to convert to. The array length indicates the number and types of values that are suggested for the method to return.</param>
-        /// <param name="parameter">The converter parameter to use.</param>
-        /// <param name="culture">The culture to use in the converter.</param>
-        /// <returns>An array of values that have been converted from the target value back to the source values.</returns>
-        public abstract object[] ConvertBack(object value, Type[] targetTypes, object parameter, CultureInfo culture);
-
-        /// <summary>
-        /// Returns an object that is provided as the value of the target property for this markup extension.
-        /// </summary>
-        /// <param name="serviceProvider">A service provider helper that can provide services for the markup extension.</param>
-        /// <returns>The object value to set on the property where the extension is applied.</returns>
-        public override object ProvideValue(IServiceProvider serviceProvider) => this;
-    }
-
     public abstract class MultiConverterBase<TSourceIn, TSourceOut, TParam, TDestination> : MultiConverterBase
     {
 #if !WinCopies3
@@ -145,7 +110,7 @@ namespace WinCopies.Util.Data
 
 #if !WinCopies3
         public static bool CheckForNullItem(in object value, in bool methodParameter) => !methodParameter && value == null;
-        
+
         private static System.Collections.Generic.IEnumerable<T> To<T>(in System.Collections.IEnumerable enumerable) => enumerable.SelectConverter(value => (T)value);
 #endif